Add files via upload
This is a basic TODO app , in which we can create our new task tickets assign priority to them using colors. Tickets can be "filtered" on the basis of colors, and their priority can be changed. There is a "unique id" associated with each ticket generated using "shortid unpkg". Tickets are stored my tickets in "local storage" hence they persists even on refreshing the tab . Tickets can be "edited" after "unlock" and can be removed  on clicking once the remove button is active